Moved container package from client to pkg
