+++ client/compat/container.go
@@ -13,10 +13,10 @@ import (
 	"github.com/go-openapi/runtime"
 	runtimeclient "github.com/go-openapi/runtime/client"
 	"github.com/skupperproject/skupper/api/types"
-	"github.com/skupperproject/skupper/client/container"
 	"github.com/skupperproject/skupper/client/generated/libpod/client/containers_compat"
 	"github.com/skupperproject/skupper/client/generated/libpod/client/exec_compat"
 	"github.com/skupperproject/skupper/client/generated/libpod/models"
+	"github.com/skupperproject/skupper/pkg/container"
 )
 
 func (c *CompatClient) ContainerList() ([]*container.Container, error) {
