+++ client/podman/container_mock.go
@@ -11,12 +11,12 @@ import (
 	"github.com/go-openapi/runtime"
 	"github.com/go-openapi/strfmt"
 	"github.com/google/uuid"
-	"github.com/skupperproject/skupper/client/container"
 	"github.com/skupperproject/skupper/client/generated/libpod/client/containers"
 	"github.com/skupperproject/skupper/client/generated/libpod/client/networks"
 	"github.com/skupperproject/skupper/client/generated/libpod/client/system"
 	"github.com/skupperproject/skupper/client/generated/libpod/client/volumes"
 	"github.com/skupperproject/skupper/client/generated/libpod/models"
+	"github.com/skupperproject/skupper/pkg/container"
 )
 
 func NewPodmanClientMock(containers []*container.Container) *PodmanRestClient {
