+++ pkg/domain/podman/link.go
@@ -3,13 +3,14 @@ package podman
 import (
 	"context"
 	"fmt"
-	"github.com/skupperproject/skupper/pkg/network"
 	"path"
 	"strconv"
 	"strings"
 
+	"github.com/skupperproject/skupper/pkg/container"
+	"github.com/skupperproject/skupper/pkg/network"
+
 	"github.com/skupperproject/skupper/api/types"
-	"github.com/skupperproject/skupper/client/container"
 	"github.com/skupperproject/skupper/client/podman"
 	"github.com/skupperproject/skupper/pkg/domain"
 	"github.com/skupperproject/skupper/pkg/qdr"
